requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
streamlit>=1.36.0
fastmcp>=0.2.0

langdetect>=1.0.9
textblob>=0.18.0
pydantic>=2.7.0
